feat(ssm): lease the mid-chunk EARLY sibling (tb - bs) alongside the tail
Gate-1 rig residual: 2/7 warm restores miss because the block-floored
match lands one block BELOW tb (chat-template gen-suffix divergence);
only the mid-chunk early capture at tb-bs can serve those turns, and it
was registered as a plain intermediate the lease did not cover — churn
evicted it.

- new is_tail_sibling entry flag + insert_tail_sibling() (no sweep; must
  follow insert_tail, whose supersede sweep now clears the session's
  previous tail AND sibling together — leased set stays <=2/session)
- lease covers is_tail || is_tail_sibling of the live session
- deadlock guard reworked: the lease binds only while an UNLEASED
  eligible candidate exists (an only-leased pool still yields a victim)
- siblings carry NO is_tail lookup gate (exact-prefix keyed, safe for
  sessionless lookers; different-nonzero-session lookers were already
  excluded by the general session filter)
- plain insert() overwrite clears the sibling flag (same rehoming
  breach vector as is_tail)
- finalize_midchunk_capture routes the early slot through the new
  insert_tail_sibling_snapshot (PrefixCache trait + radix impl)
- splits for the 500-LoC cap: snapshot_insert.rs, tests/snapshot_lease.rs

5 new unit tests; 51 snapshot tests green.
