+++ crates/spark-runtime/src/radix_tree/snapshot_insert.rs
@@ -1,0 +1,139 @@
+// SPDX-License-Identifier: AGPL-3.0-only
+
+//! Snapshot-index insert paths (plain, tail, tail-sibling). Split from
+//! `snapshot.rs` (file-size cap); same `SsmSnapshotIndex` impl.
+
+use super::snapshot::{SnapshotEntry, SsmSnapshotIndex};
+
+impl SsmSnapshotIndex {
+    pub(super) fn insert(
+        &mut self,
+        prefix_hash: u64,
+        snapshot_id: usize,
+        session_hash: u64,
+        token_count: usize,
+    ) -> Option<usize> {
+        for entry in &mut self.entries {
+            if entry.prefix_hash == prefix_hash {
+                let old = entry.snapshot_id;
+                entry.snapshot_id = snapshot_id;
+                entry.session_hash = session_hash;
+                entry.token_count = token_count;
+                // A fresh HBM save re-homes the prefix: it is resident again.
+                entry.tiered = false;
+                // A plain save re-homing this prefix is by definition NOT a
+                // tail. Without this, an overwrite could re-home another
+                // session's is_tail entry (new session_hash, is_tail still
+                // set), breaching the <=1-leased-entry-per-session invariant
+                // insert_tail's supersede sweep maintains.
+                entry.is_tail = false;
+                entry.is_tail_sibling = false;
+                self.access_counter += 1;
+                entry.last_access = self.access_counter;
+                return Some(old);
+            }
+        }
+        self.access_counter += 1;
+        self.stats.saves += 1;
+        self.entries.push(SnapshotEntry {
+            snapshot_id,
+            session_hash,
+            token_count,
+            prefix_hash,
+            last_access: self.access_counter,
+            tiered: false,
+            is_tail: false,
+            is_tail_sibling: false,
+        });
+        None
+    }
+
+    /// Insert the per-session TAIL snapshot, superseding this session's previous one.
+    /// Returns displaced snapshot_ids for the caller to free.
+    pub(super) fn insert_tail(
+        &mut self,
+        prefix_hash: u64,
+        snapshot_id: usize,
+        session_hash: u64,
+        token_count: usize,
+    ) -> Vec<usize> {
+        let mut displaced = Vec::new();
+        if session_hash != 0 {
+            let mut i = 0;
+            while i < self.entries.len() {
+                if (self.entries[i].is_tail || self.entries[i].is_tail_sibling)
+                    && self.entries[i].session_hash == session_hash
+                {
+                    displaced.push(self.entries.swap_remove(i).snapshot_id);
+                } else {
+                    i += 1;
+                }
+            }
+        }
+        for entry in &mut self.entries {
+            if entry.prefix_hash == prefix_hash {
+                displaced.push(entry.snapshot_id);
+                entry.snapshot_id = snapshot_id;
+                entry.session_hash = session_hash;
+                entry.token_count = token_count;
+                entry.is_tail = true;
+                entry.is_tail_sibling = false;
+                self.access_counter += 1;
+                entry.last_access = self.access_counter;
+                return displaced;
+            }
+        }
+        self.access_counter += 1;
+        self.entries.push(SnapshotEntry {
+            snapshot_id,
+            session_hash,
+            token_count,
+            prefix_hash,
+            last_access: self.access_counter,
+            tiered: false,
+            is_tail: true,
+            is_tail_sibling: false,
+        });
+        displaced
+    }
+
+    /// Insert the tail's EARLY sibling (`tb - bs`). MUST be called after
+    /// [`Self::insert_tail`] within the same finalize — the tail insert's
+    /// supersede sweep clears the session's previous tail AND sibling, so
+    /// this insert never needs (and must not run) its own sweep.
+    pub(super) fn insert_tail_sibling(
+        &mut self,
+        prefix_hash: u64,
+        snapshot_id: usize,
+        session_hash: u64,
+        token_count: usize,
+    ) -> Option<usize> {
+        for entry in &mut self.entries {
+            if entry.prefix_hash == prefix_hash {
+                let old = entry.snapshot_id;
+                entry.snapshot_id = snapshot_id;
+                entry.session_hash = session_hash;
+                entry.token_count = token_count;
+                entry.tiered = false;
+                entry.is_tail = false;
+                entry.is_tail_sibling = true;
+                self.access_counter += 1;
+                entry.last_access = self.access_counter;
+                return Some(old);
+            }
+        }
+        self.access_counter += 1;
+        self.stats.saves += 1;
+        self.entries.push(SnapshotEntry {
+            snapshot_id,
+            session_hash,
+            token_count,
+            prefix_hash,
+            last_access: self.access_counter,
+            tiered: false,
+            is_tail: false,
+            is_tail_sibling: true,
+        });
+        None
+    }
+}
