+++ crates/spark-runtime/src/radix_tree/tests/snapshot_index.rs
@@ -2,6 +2,9 @@
 
 use super::*;
 
+#[path = "snapshot_lease.rs"]
+mod lease;
+
 /// Build an entry with an explicit recency profile. `snapshot_id` doubles
 /// as a stable identity we assert on (independent of Vec index).
 fn entry(
@@ -18,6 +21,7 @@ fn entry(
         last_access,
         tiered: false,
         is_tail: false,
+        is_tail_sibling: false,
     }
 }
 
@@ -64,126 +68,6 @@ fn deep_tail_evicted_without_tail_protect() {
     // Victim is the OLDER entry (id 9) under pure LRU.
     let v = idx.session_aware_victim(false, false).unwrap();
     assert_eq!(idx.entries[v].snapshot_id, 9);
-}
-
-/// The lease shields the live session's `is_tail` restore point — even when
-/// it is the OLDEST entry in the pool — so the warm-turn restore survives
-/// same-session save pressure.
-#[test]
-fn tail_lease_protects_live_session_is_tail() {
-    let idx = index(vec![entry(7, 1, 8192, 100), tail_entry(9, 1, 6064, 50)], 1);
-    let v = idx.session_aware_victim(true, false).unwrap();
-    // Victim must NOT be the leased tail (id 9) despite its age.
-    assert_eq!(idx.entries[v].snapshot_id, 7);
-}
-
-/// Direct inversion of #278's off-target semantics: the live session's
-/// DEEPEST entry (the end-of-turn finish leaf, NOT a tail) is a normal
-/// eviction candidate. The old policy protected exactly this entry — which
-/// sits above `matched_tokens` and never restores — while the true restore
-/// point died (2026-07-20 rig: 0 hits with old protection on or off).
-#[test]
-fn finish_leaf_not_protected() {
-    let idx = index(
-        vec![
-            entry(7, 1, 6080, 50),      // finish leaf: deepest, oldest, NOT a tail
-            tail_entry(9, 1, 6064, 90), // true restore point
-        ],
-        1,
-    );
-    let v = idx.session_aware_victim(true, false).unwrap();
-    assert_eq!(
-        idx.entries[v].snapshot_id, 7,
-        "the deepest non-tail entry must be evictable"
-    );
-}
-
-/// The lease only shields the LIVE conversation's tail; a dormant session's
-/// tail is still evictable (correct — session-aware ranking evicts the
-/// stalest conversation first).
-#[test]
-fn dormant_session_tail_evictable() {
-    // session 2 is live; session 1 is dormant (older last_access).
-    let idx = index(
-        vec![
-            tail_entry(1, 1, 20000, 10), // dormant tail — should die first
-            entry(2, 2, 4000, 90),       // live shallow
-            tail_entry(3, 2, 12000, 95), // live tail — leased
-        ],
-        2,
-    );
-    let v = idx.session_aware_victim(true, false).unwrap();
-    assert_eq!(
-        idx.entries[v].snapshot_id, 1,
-        "stalest (dormant) session evicted first"
-    );
-}
-
-/// A pool of exactly one entry must still yield that entry as victim even
-/// when it is the leased tail — otherwise `save` can never reclaim and the
-/// cache deadlocks.
-#[test]
-fn single_leased_entry_still_evictable() {
-    let idx = index(vec![tail_entry(5, 1, 16000, 50)], 1);
-    let v = idx.session_aware_victim(true, false).unwrap();
-    assert_eq!(idx.entries[v].snapshot_id, 5);
-}
-
-/// The lease LAPSES: if the leased session never looks up again (it ended;
-/// cold churn never reaches lookup at matched_tokens == 0), its tail becomes
-/// a normal candidate after the TTL so a dead session cannot squat a slot.
-#[test]
-fn lease_expires_without_live_lookups() {
-    let mut idx = index(vec![tail_entry(9, 1, 6064, 50), entry(7, 2, 4000, 100)], 1);
-    // Lease fresh: the newer non-tail entry is the victim... but session 2 is
-    // FRESHER than session 1, so session-staleness picks session 1 first and
-    // the lease deflects to... there is only the tail in session 1, so the
-    // victim is the session-2 entry.
-    assert_eq!(idx.evict_lru(), Some(7), "leased tail survives while fresh");
-    idx.entries.push(entry(8, 2, 4000, 101));
-    // Simulate TTL expiry: many evictions with no lookup from session 1.
-    idx.evictions_since_lookup = super::tail_lease_ttl();
-    assert_eq!(
-        idx.evict_lru(),
-        Some(9),
-        "expired lease: the dead session's tail is evictable (stalest session)"
-    );
-}
-
-/// Overwriting a prefix via plain `insert` clears `is_tail` — otherwise the
-/// overwrite re-homes another session's tail (new session_hash, stale
-/// is_tail=true), breaching the <=1-leased-entry invariant and leaking tail
-/// semantics onto a non-tail save.
-#[test]
-fn insert_overwrite_clears_is_tail() {
-    let mut idx = SsmSnapshotIndex::new();
-    let displaced = idx.insert_tail(0xAB, /*slot*/ 1, /*session*/ 7, /*tok*/ 500);
-    assert!(displaced.is_empty());
-    assert!(idx.entries[0].is_tail);
-    // A plain save re-homes the same prefix for a DIFFERENT session.
-    let old = idx.insert(0xAB, /*slot*/ 2, /*session*/ 8, /*tok*/ 500);
-    assert_eq!(old, Some(1));
-    assert!(
-        !idx.entries[0].is_tail,
-        "plain insert must clear is_tail on overwrite"
-    );
-}
-
-/// The serving-path lookup (`lookup_tiered`) must session-gate `is_tail`
-/// entries exactly like the reference `lookup`: a tail is byte-safe ONLY for
-/// the same non-zero session. (The gate was previously missing here.)
-#[test]
-fn lookup_tiered_tail_session_gate() {
-    let mut idx = SsmSnapshotIndex::new();
-    let toks: Vec<u32> = (0..64).collect();
-    let ph = super::hash_token_prefix(&toks, 64, 0);
-    idx.insert_tail(ph, /*slot*/ 4, /*session*/ 7, /*tok*/ 64);
-    // Same session: hit.
-    assert!(idx.lookup_tiered(&toks, 64, 7, 0).is_some());
-    // Different session: the tail must NOT bleed.
-    assert!(idx.lookup_tiered(&toks, 64, 8, 0).is_none());
-    // Sessionless lookup: must NOT bleed either.
-    assert!(idx.lookup_tiered(&toks, 64, 0, 0).is_none());
 }
 
 // ─────────── 07-10 fossil-pinning regressions (re-landed, #317 revert) ──────────
@@ -388,44 +272,4 @@ fn reinsert_unspills() {
     idx.insert(0xAA, 5, 7, 40);
     // The entry is resident again at slot 5; the drop path can free it.
     assert_eq!(idx.evict_lru(), Some(5));
-}
-
-// ─────────────── Marconi α-score (staged, inert at default α=0) ───────────────
-
-/// At the default α=0 the within-session rank is EXACTLY pure-LRU: the
-/// oldest entry of the stalest session is the victim, depth ignored.
-#[test]
-fn alpha_zero_is_pure_lru_ordering() {
-    // Same session: deep-but-old vs shallow-but-fresh. α=0 → oldest dies.
-    let idx = index(vec![entry(1, 7, 20000, 10), entry(2, 7, 100, 90)], 7);
-    let v = idx.session_aware_victim(false, false).unwrap();
-    assert_eq!(idx.entries[v].snapshot_id, 1, "α=0 ranks purely by recency");
-}
-
-/// With α set, depth outweighs a modest recency gap WITHIN a session (a
-/// deep tail outscores a shallow fresh save), while session staleness stays
-/// the PRIMARY key — a fresher session's shallow entry still outlives a
-/// staler session entirely. α is injected (never via env — process-global
-/// mutation races the parallel test runner).
-#[test]
-fn alpha_prefers_depth_within_session_staleness_still_primary() {
-    // One session: deep old entry vs shallow slightly-fresher entry.
-    let idx = index(vec![entry(1, 7, 20000, 50), entry(2, 7, 100, 60)], 7);
-    let v = idx
-        .session_aware_victim_with_alpha(false, false, 2.0)
-        .unwrap();
-    assert_eq!(
-        idx.entries[v].snapshot_id, 2,
-        "α=2: the shallow entry is the victim despite being fresher"
-    );
-    // Two sessions: staleness first regardless of α — the stale session's
-    // DEEP entry still dies before the fresh session's shallow one.
-    let idx2 = index(vec![entry(1, 1, 20000, 10), entry(2, 2, 100, 90)], 2);
-    let v2 = idx2
-        .session_aware_victim_with_alpha(false, false, 2.0)
-        .unwrap();
-    assert_eq!(
-        idx2.entries[v2].snapshot_id, 1,
-        "session staleness remains the primary key at any α"
-    );
 }
